backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
xgboost
pydantic
langchain
langchain-community
langchain-chroma
chromadb
pypdf
pdfplumber
boto3
langchain-aws
python-dotenv
